Extract GetPublicIdentifierCallable
Extract GetPublicIdentifierCallable to it's own class.
